% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anae}
\alias{anae}
\title{Atlas of North American English Data}
\format{a data.frame with 14 columns
\describe{
 \item{speakerID}{numeric code for each speaker}
 \item{age}{speakers' age}
 \item{sex}{speakers' sex}
 \item{city}{speakers' city}
 \item{state}{speakers' state (abbreviation)}
 \item{dialect}{speakers' dialect area (ANAE definition)}
 \item{word}{word that the vowel is from}
 \item{vclass}{vowel class of measurement, in slightly modified Labovian notation}
 \item{F1}{unnormalized F1 (hz)}
 \item{F2}{unnormalized F2 (hz)}
 \item{stress}{vowel stress}
 \item{final}{whether or not the vowel was word final}
 \item{pre_seg}{the preceding segment}
 \item{fol_seg}{the following segment}
}}
\source{
Labov, W., Ash, S., & Boberg, C. (2006). The Atlas of North American English: Phonetics, Phonology and Sound Change. Change. New York: Mouton de Gruyter.
}
\usage{
anae
}
\description{
Unnormalized F1 and F2 data from the Atlas of North American English
}
\keyword{datasets}
