% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ay}
\alias{ay}
\title{/ay/ data from the Philadelphia Neighborhood Corpus}
\format{a data.frame with 16 columns
\describe{
\item{idstring}{unique id for each speaker}
\item{sex}{speakers' sex}
\item{age}{speakers' age}
\item{year}{year of interview}
\item{years_of_schooling}{speakers' years of schooling}
\item{stress}{vowel stress}
\item{word}{word}
\item{plt_vclass}{vowel class in Labovian notation}
\item{F1}{unnormalized F1}
\item{F2}{unnormalized F1}
\item{F1_n}{normalized F1 (z-scored)}
\item{F2_n}{normalized F2 (z-scored)}
\item{dur}{vowel duration}
\item{context}{location of vowel in word: coextensive, final, initial, internal}
\item{pre_seg}{preceding segment}
\item{fol_seg}{following segment}
\item{word_trans}{ARPABET transcription of the word}
}}
\source{
Labov, W., Rosenfelder, I. (2011). The Philadelphia Neighborhood Corpus of LING560 Studies

Fruehwald, J. (2016). The early influence of phonology on a phonetic change. Language, 92(2), 376–410. https://doi.org/10.1353/lan.2016.0041
}
\usage{
ay
}
\description{
The /ay/ data used in Fruehwald (2016)
}
\keyword{datasets}
