% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iy_ah}
\alias{iy_ah}
\title{/i:/ and /ɑ/ data from the Philadelphia Neighborhood Corpus}
\format{a data.frame with 17 columns
\describe{
\item{idstring}{unique id for each speaker}
\item{age}{speakers' age}
\item{sex}{speakers' sex}
\item{year}{year of interview}
\item{years_of_schooling}{speakers' years of schooling}
\item{vowel}{arpabet vowel class}
\item{word}{word}
\item{F1}{unnormalized F1}
\item{F2}{unnormalized F1}
\item{dur}{vowel duration in seconds}
\item{plt_vclass}{modifed Labovian notation of vowel class. \code{iy} = /i:/ and \code{ah} = /ɑ/}
\item{pre_seg}{preceding segment}
\item{fol_seg}{following segment}
\item{context}{location of vowel in word: coextensive, final, initial, internal}
\item{word_trans}{ARPABET transcription of the word}
\item{F1_n}{normalized F1 (z-scored)}
\item{F2_n}{normalized F2 (z-scored)}
}}
\source{
Labov, W., Rosenfelder, I. (2011). The Philadelphia Neighborhood Corpus of LING560 Studies
}
\usage{
iy_ah
}
\description{
All /i:/ and /ɑ/ data from the PNC
}
\keyword{datasets}
