% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iy_ah_wide}
\alias{iy_ah_wide}
\title{/i:/ and /ɑ/ mean data from the Philadelphia Neighborhood Corpus}
\format{a data.frame with 8 columns
\describe{
\item{idstring}{unique id for each speaker}
\item{age}{speakers' age}
\item{sex}{speakers' sex}
\item{year}{year of interview}
\item{ah_F1}{mean F1 of /ɑ/}
\item{ah_F2}{mean F2 of /ɑ/}
\item{iy_F1}{mean F1 of /i:/}
\item{iy_F2}{mean F2 of /i:/}
}}
\source{
Labov, W., Rosenfelder, I. (2011). The Philadelphia Neighborhood Corpus of LING560 Studies
}
\usage{
iy_ah_wide
}
\description{
mean /i:/ and /ɑ/ data from the PNC
}
\keyword{datasets}
