% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verboten.R
\name{verboten_vector}
\alias{verboten_vector}
\title{Create Verboten Vector}
\usage{
verboten_vector(x = c("PH76-3-1-", "PH84-2-5-", "PH84-2-3-", "PH81-1-9-",
  "PH92-1-1-", "PH82-1-16-", "PH80-2-5-", "PH80-2-4-", "PH85-1-5-", "PH84-1-2-",
  "PH84-1-1-", "PH77-3-4-"))
}
\description{
vector of non-Philadelphian speakers
}
